Rename ResearchOutputTeamForm to ResearchOutputForm
