+++ packages/react-components/src/templates/index.ts
@@ -40,7 +40,7 @@ export { default as OpenQuestionsModal } from './OpenQuestionsModal';
 export { default as PasswordResetEmailSentPage } from './PasswordResetEmailSentPage';
 export { default as PersonalInfoModal } from './PersonalInfoModal';
 export { default as ProfileOutputs } from './ProfileOutputs';
-export { default as ResearchOutputTeamForm } from './ResearchOutputTeamForm';
+export { default as ResearchOutputForm } from './ResearchOutputForm';
 export { default as ResearchOutputsSearch } from './ResearchOutputsSearch';
 export { default as RoleModal } from './RoleModal';
 export { default as SharedResearchList } from './SharedResearchList';
